gnn/requirements.txt
# torch==2.4.0+cu124
# dgl==2.4.0+cu124
torchdata==0.7.1
pandas==2.2.2
pydantic
transformers==4.38.0